+++ Microsoft.Azure.Cosmos/src/ChangeFeedProcessor/FeedProcessing/FeedEstimatorCore.cs
@@ -8,7 +8,7 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.FeedProcessing
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos.ChangeFeed.FeedManagement;
-    using Microsoft.Azure.Documents;
+    using Microsoft.Azure.Cosmos.Core.Trace;
 
     internal sealed class FeedEstimatorCore : FeedEstimator
     {
@@ -40,7 +40,7 @@ namespace Microsoft.Azure.Cosmos.ChangeFeed.FeedProcessing
                     if (cancellationToken.IsCancellationRequested)
                         throw;
 
-                    DefaultTrace.TraceException(new Exception("exception within estimator", canceledException));
+                    Extensions.TraceException(new Exception("exception within estimator", canceledException));
 
                     // ignore as it is caused by client
                 }