+++ Microsoft.Azure.Cosmos/src/Routing/ClientCollectionCache.cs
@@ -64,7 +64,7 @@ namespace Microsoft.Azure.Cosmos.Routing
                    ResourceType.Collection,
                    collectionLink,
                    AuthorizationTokenType.PrimaryMasterKey,
-                   new StringKeyValueCollection()))
+                   new DictionaryNameValueCollection()))
             {
                 request.Headers[HttpConstants.HttpHeaders.XDate] = DateTime.UtcNow.ToString("r");