+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.EmulatorTests/QueryTests.cs
@@ -469,7 +469,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                         NumericField2 = index
                     };
 
-                    INameValueCollection headers = new StringKeyValueCollection();
+                    INameValueCollection headers = new DictionaryNameValueCollection();
                     if (!collection.IndexingPolicy.Automatic)
                     {
                         headers.Add("x-ms-indexing-directive", "include");
@@ -601,7 +601,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                 doc.StringField = "222";
                 Document documentDefinition = (Document)doc;
                 documentDefinition.SetPropertyValue("pk", "test");
-                INameValueCollection requestHeaders = new StringKeyValueCollection();
+                INameValueCollection requestHeaders = new DictionaryNameValueCollection();
                 requestHeaders.Add("x-ms-indexing-directive", "exclude");
                 this.client.Create<Document>(collection.GetIdOrFullName(), documentDefinition, requestHeaders);
 
@@ -637,7 +637,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                     StringField = "333",
                 };
                 doc.SetPropertyValue("pk", "test");
-                INameValueCollection requestHeaders = new StringKeyValueCollection();
+                INameValueCollection requestHeaders = new DictionaryNameValueCollection();
                 requestHeaders.Add("x-ms-indexing-directive", "include");
 
                 QueryDocument docCreated = this.client.Create<QueryDocument>(collection.GetIdOrFullName(), doc, requestHeaders);
@@ -790,7 +790,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                     sourceCollection.IndexingPolicy.IndexingMode = IndexingMode.Consistent;
                     DocumentCollection collection = testClient.Create<DocumentCollection>(database.GetIdOrFullName(), sourceCollection);
 
-                    INameValueCollection requestHeaders = new StringKeyValueCollection();
+                    INameValueCollection requestHeaders = new DictionaryNameValueCollection();
                     requestHeaders.Add("x-ms-indexing-directive", "include");
 
                     Action<string, string, string> testDocumentSQL = (name, rawValue, escapedValue) =>
@@ -1669,7 +1669,7 @@ namespace Microsoft.Azure.Cosmos.SDK.EmulatorTests
                 SqlQuerySpec querySpec = new SqlQuerySpec(string.Format("SELECT * FROM r"));
                 using (HttpClient httpClient = new HttpClient())
                 {
-                    var headers = new StringKeyValueCollection();
+                    var headers = new DictionaryNameValueCollection();
                     httpClient.AddMasterAuthorizationHeader("post", coll.ResourceId, "docs", headers, masterKey);
                     httpClient.DefaultRequestHeaders.Add(HttpConstants.HttpHeaders.IsQuery, bool.TrueString);
                     httpClient.DefaultRequestHeaders.Add(HttpConstants.HttpHeaders.EnableScanInQuery, bool.TrueString);
@@ -2743,7 +2743,7 @@ function sproc(feed) {
                     StringField = index.ToString(CultureInfo.InvariantCulture),
                 };
                 doc.SetPropertyValue("pk", "test");
-                INameValueCollection headers = new StringKeyValueCollection();
+                INameValueCollection headers = new DictionaryNameValueCollection();
                 if (!collection.IndexingPolicy.Automatic && manualIndex)
                 {
                     headers.Add("x-ms-indexing-directive", "include");
@@ -2814,7 +2814,7 @@ function sproc(feed) {
                     StringField = index.ToString(CultureInfo.InvariantCulture),
                 };
 
-                INameValueCollection headers = new StringKeyValueCollection();
+                INameValueCollection headers = new DictionaryNameValueCollection();
                 if (!collection.IndexingPolicy.Automatic && manualIndex)
                 {
                     headers.Add("x-ms-indexing-directive", "include");